+++ src/main/java/org/sopt/snappinserver/domain/product/service/usecase/GetProductReviewsUseCase.java
@@ -1,4 +1,4 @@
-package org.sopt.snappinserver.domain.product.service;
+package org.sopt.snappinserver.domain.product.service.usecase;
 
 import org.sopt.snappinserver.domain.review.service.dto.response.ReviewPageResult;
